refactor(adapter-ui,adapter-server): extract action-error helper + English-only comment
Round 3 review (coderabbit):
- resolveActionErrorMessage (this PR's addition to api.ts) moves to a
  new focused module src/lib/action-errors.ts; all importers updated,
  no re-export shim. api.ts shrinks to 1395 lines — smaller than before
  this PR touched it. The full api.ts split stays tracked separately.
- the rule-block sanitization-exception comment now uses an
  English-only example per the repo's comments-in-English rule
