+++ addons/adapter-ui/cap-adapter-ui/src/pages/entity-form-actions.ts
@@ -14,13 +14,13 @@ import type { EnrichedSubmitData } from "../components/auto-form/types";
 import type { StateTransitionInfo } from "../components/status-bar";
 import type { ResolvedEntityBundle } from "../hooks/use-entity-bundle";
 import { pushNotification } from "../hooks/use-notifications";
+import { resolveActionErrorMessage } from "../lib/action-errors";
 import {
   createRecord,
   deleteRecord,
   executeAction,
   queryRecord,
   queryStateTransitions,
-  resolveActionErrorMessage,
   updateRecord,
 } from "../lib/api";
 import { CLONE_STRIP_FIELDS, getMutationReturnFields } from "../lib/entity-form-utils";
